adjusted WheeledRobotClass to use actions, compile tested, ready to be practicaly tested

+++ include/UGVNavigator.hpp
@@ -41,7 +41,7 @@ class UGVNavigator : public msg_emitter, public msg_receiver, public TimedBlock 
         Timer m_Timer;
         Line2D m_FireDirection;
         UGVPatrolState m_PatrolState = UGVPatrolState::IDLE;
-        Vector2D<double> m_HomePosition, m_EntrancePosition, m_FireLocation, m_ExtLocation;
+        Vector2D<float> m_HomePosition, m_EntrancePosition, m_FireLocation, m_ExtLocation;
         int m_SearchTimeOut = 5000, m_ReachingGoalPositionTimeOut = 40000;
         float m_HomeHeading, m_EntranceHeading, m_FireHeading, m_ExtHeading;
 };
